requirements.txt
fastapi==0.104.1
google-generativeai>=0.8.4
python-dotenv==1.0.0
pydantic==2.5.0

